Rename web3_registry.py
